Fix matchrange

+++ src/Regex0.cs
@@ -351,7 +351,7 @@ namespace System.Text.RegularExpressions.RegexLight
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static bool matchrange(char c, ReadOnlySpan<char> str)
         {
-            return (str.Length >= 3) && (str[1] == '-') && ((uint)(c - str[0]) <= str[2]);
+            return (str.Length >= 3) && (str[1] == '-') && ((uint)(c - str[0]) <= (uint)str[2]);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
